admin: refacto information section into read component
