+++ admin/app/components/organizations/information-section.hbs
@@ -153,104 +153,11 @@
       </div>
 
     {{else}}
-      <div class="organization__data">
-        <h1 class="organization__name">{{@organization.name}}</h1>
-
-        {{#if @organization.tags}}
-          <ul class="organization-tags-list">
-            {{#each @organization.tags as |tag|}}
-              <li class="organization-tags-list__tag">
-                <PixTag @compact="true" @color="purple-light">{{tag.name}}</PixTag>
-              </li>
-            {{/each}}
-          </ul>
-        {{/if}}
-
-        {{#if @organization.isArchived}}
-          <PixMessage class="organization-information-section__archived-message" @type="warning">
-            Archivée le
-            {{@organization.archivedFormattedDate}}
-            par
-            {{@organization.archivistFullName}}.
-          </PixMessage>
-        {{/if}}
-
-        <div class="organization-information-section__content">
-          <div class="organization-information-section__details">
-            <ul>
-              <li>Type : {{@organization.type}}</li>
-              <li>Créée par : {{@organization.creatorFullName}} ({{@organization.createdBy}})</li>
-              {{#if @organization.externalId}}
-                <li>Identifiant externe : {{@organization.externalId}}</li>
-              {{/if}}
-              {{#if @organization.provinceCode}}
-                <li>Département : {{@organization.provinceCode}}</li>
-              {{/if}}
-              {{#if (or @organization.isOrganizationSCO @organization.isOrganizationSUP)}}
-                <li>Gestion d’élèves/étudiants : {{if @organization.isManagingStudents "Oui" "Non"}}</li>
-              {{/if}}
-              <li>Adresse e-mail d'activation SCO : {{@organization.email}}</li>
-              <li>Affichage des acquis dans l'export de résultats : {{if @organization.showSkills "Oui" "Non"}}</li>
-              <li>Affichage du Net Promoter Score : {{if @organization.showNPS "Oui" "Non"}}</li>
-              <li>Lien vers le formulaire du Net Promoter Score :
-                {{#if @organization.formNPSUrl}}
-                  <a
-                    href="{{@organization.formNPSUrl}}"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                  >{{@organization.formNPSUrl}}</a>
-                {{else}}
-                  Non spécifié
-                {{/if}}
-              </li>
-              <li>Crédits : {{@organization.credit}}</li>
-              <li>Lien vers la documentation :
-                {{#if @organization.documentationUrl}}
-                  <a
-                    href="{{@organization.documentationUrl}}"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                  >{{@organization.documentationUrl}}</a>
-                {{else}}
-                  Non spécifié
-                {{/if}}
-              </li>
-            </ul>
-            {{#if this.accessControl.hasAccessToOrganizationActionsScope}}
-              <div class="form-actions">
-                <PixButton
-                  @backgroundColor="transparent-light"
-                  @isBorderVisible={{true}}
-                  @size="small"
-                  @triggerAction={{this.toggleEditMode}}
-                >
-                  Éditer
-                </PixButton>
-                {{#unless @organization.isArchived}}
-                  <PixButton
-                    @backgroundColor="red"
-                    @size="small"
-                    @triggerAction={{this.toggleArchivingConfirmationModal}}
-                  >
-                    Archiver l'organisation
-                  </PixButton>
-                {{/unless}}
-              </div>
-            {{/if}}
-          </div>
-          <div>
-            <PixButtonLink
-              @backgroundColor="transparent-light"
-              @isBorderVisible={{true}}
-              @href={{this.externalURL}}
-              @size="small"
-              class="pix-button__secondary"
-              target="_blank"
-              rel="noopener noreferrer"
-            >Tableau de bord</PixButtonLink>
-          </div>
-        </div>
-      </div>
+      <Organizations::Information-section-read-mode
+        @organization={{@organization}}
+        @toggleEditMode={{this.toggleEditMode}}
+        @toggleArchivingConfirmationModal={{this.toggleArchivingConfirmationModal}}
+      />
     {{/if}}
 
     {{#if this.showArchivingConfirmationModal}}
