+++ admin/tests/integration/components/organizations/information-section-read-mode_test.js
@@ -1,0 +1,185 @@
+import { module, test } from 'qunit';
+import { setupRenderingTest } from 'ember-qunit';
+import { render } from '@1024pix/ember-testing-library';
+import hbs from 'htmlbars-inline-precompile';
+import EmberObject from '@ember/object';
+import Service from '@ember/service';
+
+module('Integration | Component | organizations/information-section-read-mode', function (hooks) {
+  setupRenderingTest(hooks);
+
+  module('when user has access', function (hooks) {
+    hooks.beforeEach(function () {
+      class AccessControlStub extends Service {
+        hasAccessToOrganizationActionsScope = true;
+      }
+      this.owner.register('service:access-control', AccessControlStub);
+    });
+
+    test('it renders general information about organization', async function (assert) {
+      // given
+      this.set('organization', {
+        type: 'SUP',
+        isManagingStudents: false,
+        name: 'SUPer Orga',
+        credit: 350,
+        documentationUrl: 'https://pix.fr',
+        showSkills: true,
+        createdBy: 1,
+        creatorFullName: 'Gilles Parbal',
+      });
+
+      // when
+      const screen = await render(
+        hbs`<Organizations::InformationSectionReadMode @organization={{this.organization}} />`
+      );
+
+      // then
+      assert.dom(screen.getByRole('heading', { name: 'SUPer Orga' })).exists();
+      assert.dom(screen.getByText('Type : SUP')).exists();
+      assert.dom(screen.getByText('Créée par : Gilles Parbal (1)')).exists();
+      assert.dom(screen.getByText("Affichage des acquis dans l'export de résultats : Oui")).exists();
+      assert.dom(screen.getByText('Crédits : 350')).exists();
+      assert.dom(screen.getByText('https://pix.fr')).exists();
+    });
+
+    test('it renders edit and archive button', async function (assert) {
+      // given
+      this.organization = EmberObject.create({ type: 'SUP', isManagingStudents: false, name: 'SUPer Orga' });
+
+      // when
+      const screen = await render(
+        hbs`<Organizations::InformationSectionReadMode @organization={{this.organization}} />`
+      );
+
+      // then
+      assert.dom(screen.getByRole('button', { name: 'Éditer' })).exists();
+      assert.dom(screen.getByRole('button', { name: "Archiver l'organisation" })).exists();
+    });
+
+    test('it should display empty documentation link message', async function (assert) {
+      // given
+      const organization = EmberObject.create({});
+      this.set('organization', organization);
+
+      // when
+      const screen = await render(
+        hbs`<Organizations::InformationSectionReadMode @organization={{this.organization}} />`
+      );
+
+      // then
+      assert.dom(screen.getByText('Lien vers la documentation : Non spécifié')).exists();
+    });
+
+    test('it should display tags', async function (assert) {
+      // given
+      const organization = EmberObject.create({
+        tags: [
+          { id: 1, name: 'CFA' },
+          { id: 2, name: 'PRIVE' },
+          { id: 3, name: 'AGRICULTURE' },
+        ],
+      });
+      this.set('organization', organization);
+
+      // when
+      const screen = await render(
+        hbs`<Organizations::InformationSectionReadMode @organization={{this.organization}} />`
+      );
+
+      // expect
+      assert.dom(screen.getByText('CFA')).exists();
+      assert.dom(screen.getByText('PRIVE')).exists();
+      assert.dom(screen.getByText('AGRICULTURE')).exists();
+    });
+
+    module('when organization is archived', function () {
+      test('it should display who archived it', async function (assert) {
+        // given
+        const store = this.owner.lookup('service:store');
+        const archivedAt = new Date('2022-02-22');
+        const organization = store.createRecord('organization', { archivistFullName: 'Rob Lochon', archivedAt });
+        this.set('organization', organization);
+
+        // when
+        const screen = await render(
+          hbs`<Organizations::InformationSectionReadMode @organization={{this.organization}} />`
+        );
+
+        // expect
+        assert.dom(screen.getByText('Archivée le 22/02/2022 par Rob Lochon.')).exists();
+      });
+    });
+
+    module('When organization is SCO or SUP', function (hooks) {
+      hooks.beforeEach(function () {
+        this.organization = EmberObject.create({ type: 'SCO', isOrganizationSCO: true, isManagingStudents: true });
+      });
+
+      test('it should display "Oui" if it is managing students', async function (assert) {
+        // given
+        this.organization.isManagingStudents = true;
+
+        // when
+        const screen = await render(
+          hbs`<Organizations::InformationSectionReadMode @organization={{this.organization}} />`
+        );
+
+        assert.dom(screen.getByText(`Gestion d’élèves/étudiants : Oui`)).exists();
+      });
+
+      test('it should display "Non" if managing students is false', async function (assert) {
+        // given
+        this.organization.isManagingStudents = false;
+
+        // when
+        const screen = await render(
+          hbs`<Organizations::InformationSectionReadMode @organization={{this.organization}} />`
+        );
+
+        // then
+        assert.dom(screen.getByText(`Gestion d’élèves/étudiants : Non`)).exists();
+      });
+    });
+
+    module('When organization is not SCO', function (hooks) {
+      hooks.beforeEach(function () {
+        this.organization = EmberObject.create({ type: 'PRO', isOrganizationSCO: false, isOrganizationSUP: false });
+      });
+
+      test('it should not display if it is managing students', async function (assert) {
+        // given
+        this.organization.isManagingStudents = false;
+
+        // when
+        const screen = await render(
+          hbs`<Organizations::InformationSectionReadMode @organization={{this.organization}} />`
+        );
+
+        // then
+        assert.dom(screen.queryByRole('checkbox', { name: 'Gestion d’élèves/étudiants' })).doesNotExist();
+      });
+    });
+  });
+
+  module('when user does not have access', function () {
+    test('it should not allow to edit or archive an organization', async function (assert) {
+      // given
+      const organization = EmberObject.create({ name: 'Boba Fett' });
+      this.set('organization', organization);
+      class AccessControlStub extends Service {
+        hasAccessToOrganizationActionsScope = false;
+      }
+      this.owner.register('service:access-control', AccessControlStub);
+
+      // when
+      const screen = await render(
+        hbs`<Organizations::InformationSectionReadMode @organization={{this.organization}} />`
+      );
+
+      // expect
+      assert.dom(screen.queryByRole('button', { name: 'Éditer' })).doesNotExist();
+      assert.dom(screen.queryByRole('button', { name: "Archiver l'organisation" })).doesNotExist();
+    });
+  });
+});
