+++ admin/tests/unit/components/organizations/information-section_test.js
@@ -8,7 +8,7 @@ module('Unit | Component | organizations/information-section', function (hooks) 
 
   test('it should generate link based on environment and object', async function (assert) {
     // given
-    const component = createGlimmerComponent('component:organizations/information-section');
+    const component = createGlimmerComponent('component:organizations/information-section-read-mode');
     ENV.APP.ORGANIZATION_DASHBOARD_URL = 'https://metabase.pix.fr/dashboard/137/?id=';
     component.args = { organization: { id: 1 } };
